Moved the categories page to the index page.
