+++ src/components/hero.module.css
@@ -15,6 +15,8 @@
 }
 
 .heroImage {
+  padding-left: 1rem;
+  padding-right: 1rem;
   padding-top: 2rem;
   padding-bottom: 2rem;
   height: auto;
