+++ src/pages/index.module.css
@@ -1,0 +1,29 @@
+.button {
+    border-color: var(--text-accent);
+    border-width: 6px;
+    border-style: solid;
+
+    display:block;
+    text-align:center;
+    padding: 1rem;
+    margin-top: 1rem;
+    font-size: 18px;
+    transition: all 200ms ease;
+  }
+  
+  .button:hover, .button:active {
+    border-color: var(--text-dark);
+    text-decoration: none;
+  }
+
+/**
+ * article grid
+ */
+.list {
+  margin: 0;
+  padding: 0;
+  list-style: none;
+  display: grid;
+  grid-template-columns: repeat(auto-fit, minmax(300px, 1fr));
+  grid-gap: 5vmin;
+}
